move rails specific bits to spec_helper_rails
this allows you to typically only include spec_helper for rspec tests which do not rely on rails (e.g., most test in the lib folder).

'spec_helper' does not load up rails and incur the typically overhead of loading rails. If you have a test that requires rails, it should include 'spec_helper_rails'

This flow allows for fast tests recoupled from rails while still allowing tests to depend on rails. As a note, most things in lib should work to not use rails as a dependency